+++ backend/database/share/share.go
@@ -48,15 +48,17 @@ type CreateBody struct {
 // Link is the information needed to build a shareable link.
 type Link struct {
 	CommonShare
-	Mu            sync.Mutex     `json:"-"`
-	Downloads     int            `json:"downloads"`
-	UserDownloads map[string]int `json:"-"` // Track downloads per username (not persisted)
-	Hash          string         `json:"hash" storm:"id,index"`
-	UserID        uint           `json:"userID"`
-	Expire        int64          `json:"expire"`
-	PasswordHash  string         `json:"password_hash,omitempty"`
+	Downloads    int    `json:"downloads"`
+	Hash         string `json:"hash" storm:"id,index"`
+	UserID       uint   `json:"userID"`
+	Expire       int64  `json:"expire"`
+	PasswordHash string `json:"password_hash,omitempty"`
 	// Token is a random value that will only be set when PasswordHash is set. It is
 	// URL-Safe and is used to download links in password-protected shares via a
 	// query arg.
 	Token string `json:"token,omitempty"`
+
+	Mu            sync.Mutex     `json:"-"`
+	UserDownloads map[string]int `json:"-"` // Track downloads per username (not persisted)
+
 }