+++ backend/http/httpRouter.go
@@ -140,6 +140,9 @@ func StartHttp(ctx context.Context, storage *bolt.BoltStore, shutdownComplete ch
 	publicAPI.HandleFunc("GET /onlyoffice/callback", withHashFile(onlyofficeCallbackHandler))
 	publicAPI.HandleFunc("GET /onlyoffice/config", withHashFile(onlyofficeClientConfigGetHandler))
 	publicAPI.HandleFunc("POST /resources", withHashFile(publicUploadHandler))
+	publicAPI.HandleFunc("PUT /resources", withHashFile(publicPutHandler))
+	publicAPI.HandleFunc("DELETE /resources", withHashFile(publicDeleteHandler))
+	publicAPI.HandleFunc("PATCH /resources", withHashFile(publicPatchHandler))
 
 	// Settings routes
 	api.HandleFunc("GET /settings", withAdmin(settingsGetHandler))